requirements.txt
numpy==1.17.2
torch==1.5.1
torchaudio==0.5.1
librosa==0.7.2
soundfile==0.10.3
tensorflow==1.15.2
scipy==1.3.1
ffmpeg-python
